add numpy.uint16 support
